+++ protocols/kad/CHANGELOG.md
@@ -6,8 +6,11 @@
 
 - Derive `Clone` for `KademliaEvent` (see [PR 2411])
 
+- Derive `Serialize`, `Deserialize` for `store::record::Key` (see [PR 2408])
+
 [PR 2339]: https://github.com/libp2p/rust-libp2p/pull/2339
 [PR 2411]: https://github.com/libp2p/rust-libp2p/pull/2411
+[PR 2408]: https://github.com/libp2p/rust-libp2p/pull/2408
 
 # 0.33.0 [2021-11-16]